Add JSON example file
I have noticed that the project is using CSV file to store data. However it is not easy to read and modify. JSON is more ideal for this kind of purpose, it has following advantages:
- Self descriptive, easy to read and modify
- Widely supported, convenient for related developments

I apologize that I didn't commit a JSON file containing all institutes recorded by this project. I'm currently unable to use computer to write a script to convert the CSV file into a JSON file. So I committed a example JSON file.